Still tweaking these joints
Character Joint is being stubborn, but honestly is close enough. Figured out the gizmo is connected to a few field that are available in the editor, that's making a little more sense now. But not a lot, so I started playing around with "hinge joint" and now I'm falling asleep so I can't assess whether it's good or bad.
Pretty aggravating to still be working on this days later when we still have no "game" in our game. (No objectives, no environment).
I think what I need to do is implement the character joint component, get it close enough to good and move on.